Change price to int as it was in original libraries.

+++ app/code/src/Payment/PaymentProcessorInterface.php
@@ -4,5 +4,5 @@ namespace App\Payment;
 
 interface PaymentProcessorInterface
 {
-    public function pay(float $price): void;
+    public function pay(int $price): void;
 }
